requirements.txt
bosdyn-api==5.0.0
bosdyn-choreography-client==5.0.0
bosdyn-client==5.0.0
bosdyn-core==5.0.0
bosdyn-mission==5.0.0
bosdyn-orbit==5.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
fonttools==4.58.2
grpcio==1.73.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.0
packaging==25.0
pillow==11.2.1
protobuf==6.31.1
PyJWT==2.10.1
pynmea2==1.19.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==80.9.0
six==1.17.0
urllib3==2.4.0
vtk==9.4.2
windows-curses==2.4.1
wrapt==1.17.2